Allow infinite number of iterations
The task now limits the number of generated records only when the `iterations` configuration parameter is set to 1 or more. This maintains the existing behavior, but allows for infinite number of records when `iterations=-1` (or really any value less than 1).

Also cleaned up the code a bit to be a little more robust and efficient, and eliminated a checkstyle warning.