Infer key schemas from value schemas, and change default key schema and value
The key schema is hard-coded as STRING_SCHEMA, specifically non-optional.
This makes it impossible to specify `null` as a key, to trigger round-robin/sticky partitioning.
It also forces the key field to always refer to a string field, and not a more complex key.

This change infers the schemas from the key field itself, or defaults to
OPTIONAL_STRING_SCHEMA schema and null value for the key if no field is specified.

Signed-off-by: Greg Harris <gregh@confluent.io>